+++ src/views/layouts/listItem.blade.php
@@ -44,11 +44,11 @@
             !!}
             {{-- Last message body --}}
             @if($lastMessage->attachment == null)
-            {{
+            {!!
                 strlen($lastMessage->body) > 30
                 ? trim(substr($lastMessage->body, 0, 30)).'..'
                 : $lastMessage->body
-            }}
+            !!}
             @else
             <span class="fas fa-file"></span> Attachment
             @endif